refactor : make heading component separately
